Performed some cleanup and added the hostname, nbns, and trigger parameters
Performed some code cleanup. Added the 'hostname' parameter for setting
the hostname to spoof with NBNS. Also added the NBNS parameter to
enable/disable NBNS. If disabled, another method of spoofing will be
needed such as an LLMNR/NBNS spoofer on another system. Finally, added
the trigger parameter to enable/disable the defender signature updates.
If disabled, another method will need to be employed to trigger the
relay.